Factored out most of the brain-related stuff into brain.py
